ansily-compliant integer (de)serialisation

+++ include/common.h
@@ -35,7 +35,7 @@ typedef int8_t s8;
 typedef int16_t s16;
 typedef int32_t s32;
 
-static s32 read_neutral_s32(u8 * data) { return data[0] | (data[1] << 8) | (data[2] << 16) | (data[3] << 24); }
+static s32 read_neutral_s32(u8 * data) { return ((u32) data[0]) | (((u32) data[1]) << 8) | (((u32) data[2]) << 16) | (((u32) data[3]) << 24); }
 
 static void write_neutral_s32(u8 * data, s32 value) {
     data[0] = value & 0xFF;
